chore: merge network store into state store
