+++ crates/agglayer-certificate-orchestrator/src/lib.rs
@@ -13,9 +13,8 @@ use agglayer_storage::{
         latest_settled_certificate_per_network::SettledCertificate,
     },
     stores::{
-        EpochStoreReader, EpochStoreWriter, LocalNetworkStateReader, LocalNetworkStateWriter,
-        PendingCertificateReader, PendingCertificateWriter, PerEpochReader, PerEpochWriter,
-        StateReader, StateWriter,
+        EpochStoreReader, EpochStoreWriter, PendingCertificateReader, PendingCertificateWriter,
+        PerEpochReader, PerEpochWriter, StateReader, StateWriter,
     },
 };
 use agglayer_types::{CertificateId, CertificateIndex, Height, NetworkId};
@@ -68,7 +67,6 @@ pub struct CertificateOrchestrator<
     EpochsStore,
     PerEpochStore,
     StateStore,
-    NetworkStateStore,
 > {
     /// Epoch packing task resolver.
     epoch_packing_tasks: EpochPackingTasks,
@@ -87,8 +85,6 @@ pub struct CertificateOrchestrator<
     /// Cancellation token for graceful shutdown.
     cancellation_token: CancellationToken,
 
-    /// Store for the local network state of each chain.
-    network_state_store: Arc<NetworkStateStore>,
     /// The state store to access data.
     state_store: Arc<StateStore>,
     /// Pending store to access the certificates and proofs.
@@ -120,15 +116,7 @@ pub struct CertificateOrchestrator<
         mpsc::Sender<(oneshot::Sender<SettledCertificate>, ProvenCertificate)>,
 }
 
-impl<
-        E,
-        CertifierClient,
-        PendingStore,
-        EpochsStore,
-        PerEpochStore,
-        StateStore,
-        NetworkStateStore,
-    >
+impl<E, CertifierClient, PendingStore, EpochsStore, PerEpochStore, StateStore>
     CertificateOrchestrator<
         E,
         CertifierClient,
@@ -136,7 +124,6 @@ impl<
         EpochsStore,
         PerEpochStore,
         StateStore,
-        NetworkStateStore,
     >
 where
     PendingStore: PendingCertificateReader,
@@ -155,7 +142,6 @@ where
         epochs_store: Arc<EpochsStore>,
         current_epoch: ArcSwap<PerEpochStore>,
         state_store: Arc<StateStore>,
-        network_state_store: Arc<NetworkStateStore>,
     ) -> Result<Self, Error> {
         let (certification_notification_sender, certification_notification) =
             mpsc::channel(Self::DEFAULT_CERTIFICATION_NOTIFICATION_CHANNEL_SIZE);
@@ -182,21 +168,12 @@ where
             settlement_notifier: Default::default(),
             certification_notification,
             certification_notification_sender,
-            network_state_store,
         })
     }
 }
 
 #[buildstructor::buildstructor]
-impl<
-        E,
-        CertifierClient,
-        PendingStore,
-        EpochsStore,
-        PerEpochStore,
-        StateStore,
-        NetworkStateStore,
-    >
+impl<E, CertifierClient, PendingStore, EpochsStore, PerEpochStore, StateStore>
     CertificateOrchestrator<
         E,
         CertifierClient,
@@ -204,7 +181,6 @@ impl<
         EpochsStore,
         PerEpochStore,
         StateStore,
-        NetworkStateStore,
     >
 where
     CertifierClient: Certifier,
@@ -213,7 +189,6 @@ where
     EpochsStore: EpochStoreWriter<PerEpochStore = PerEpochStore> + EpochStoreReader + 'static,
     PerEpochStore: PerEpochWriter + PerEpochReader + 'static,
     StateStore: StateReader + StateWriter + 'static,
-    NetworkStateStore: LocalNetworkStateReader + LocalNetworkStateWriter + 'static,
 {
     /// Function that setups and starts the CertificateOrchestrator.
     ///
@@ -244,7 +219,6 @@ where
         epochs_store: Arc<EpochsStore>,
         current_epoch: ArcSwap<PerEpochStore>,
         state_store: Arc<StateStore>,
-        network_state_store: Arc<NetworkStateStore>,
     ) -> anyhow::Result<JoinHandle<()>> {
         let mut orchestrator = Self::try_new(
             clock,
@@ -256,7 +230,6 @@ where
             epochs_store,
             current_epoch,
             state_store,
-            network_state_store,
         )?;
 
         // Try to spawn the certifier tasks for the next height of each network
@@ -272,15 +245,7 @@ where
     }
 }
 
-impl<
-        E,
-        CertifierClient,
-        PendingStore,
-        EpochsStore,
-        PerEpochStore,
-        StateStore,
-        NetworkStateStore,
-    >
+impl<E, CertifierClient, PendingStore, EpochsStore, PerEpochStore, StateStore>
     CertificateOrchestrator<
         E,
         CertifierClient,
@@ -288,7 +253,6 @@ impl<
         EpochsStore,
         PerEpochStore,
         StateStore,
-        NetworkStateStore,
     >
 where
     CertifierClient: Certifier,
@@ -297,7 +261,6 @@ where
     EpochsStore: EpochStoreWriter<PerEpochStore = PerEpochStore> + EpochStoreReader + 'static,
     StateStore: StateReader + StateWriter + 'static,
     PerEpochStore: PerEpochWriter + PerEpochReader + 'static,
-    NetworkStateStore: LocalNetworkStateReader + LocalNetworkStateWriter + 'static,
 {
     fn spawn_network_task(&mut self, network_id: NetworkId) -> Result<(), Error> {
         if self.spawned_network_tasks.contains_key(&network_id) {
@@ -311,7 +274,6 @@ where
         let task = NetworkTask::new(
             self.pending_store.clone(),
             self.state_store.clone(),
-            self.network_state_store.clone(),
             self.certifier_task_builder.clone(),
             self.certification_notification_sender.clone(),
             self.clock_ref.clone(),
@@ -441,15 +403,7 @@ where
 //
 // When a Certificate is proven, we try to add it to the current epoch.
 // If we succeed, we try to settle it on L1.
-impl<
-        E,
-        CertifierClient,
-        PendingStore,
-        EpochsStore,
-        PerEpochStore,
-        StateStore,
-        NetworkStateStore,
-    >
+impl<E, CertifierClient, PendingStore, EpochsStore, PerEpochStore, StateStore>
     CertificateOrchestrator<
         E,
         CertifierClient,
@@ -457,7 +411,6 @@ impl<
         EpochsStore,
         PerEpochStore,
         StateStore,
-        NetworkStateStore,
     >
 where
     CertifierClient: Certifier,
@@ -465,7 +418,6 @@ where
     PendingStore: PendingCertificateReader + PendingCertificateWriter,
     StateStore: StateReader + StateWriter,
     PerEpochStore: PerEpochWriter + PerEpochReader + 'static,
-    NetworkStateStore: LocalNetworkStateReader + LocalNetworkStateWriter + 'static,
 {
     fn handle_settlement_result(
         &mut self,
@@ -582,16 +534,8 @@ where
     }
 }
 
-impl<E, A, PendingStore, EpochsStore, PerEpochStore, StateStore, NetworkStateStore> Future
-    for CertificateOrchestrator<
-        E,
-        A,
-        PendingStore,
-        EpochsStore,
-        PerEpochStore,
-        StateStore,
-        NetworkStateStore,
-    >
+impl<E, A, PendingStore, EpochsStore, PerEpochStore, StateStore> Future
+    for CertificateOrchestrator<E, A, PendingStore, EpochsStore, PerEpochStore, StateStore>
 where
     A: Certifier,
     E: EpochPacker<PerEpochStore = PerEpochStore>,
@@ -599,7 +543,6 @@ where
     EpochsStore: EpochStoreWriter<PerEpochStore = PerEpochStore> + EpochStoreReader + 'static,
     StateStore: StateReader + StateWriter + 'static,
     PerEpochStore: PerEpochWriter + PerEpochReader + 'static,
-    NetworkStateStore: LocalNetworkStateReader + LocalNetworkStateWriter + 'static,
 {
     type Output = ();
