+++ crates/agglayer-storage/src/storage/cf_definitions.rs
@@ -1,7 +1,6 @@
 use rocksdb::ColumnFamilyDescriptor;
 
 pub mod epochs;
-pub mod local_network_state;
 pub mod pending;
 pub mod state;
