+++ crates/agglayer-storage/src/stores/interfaces/reader.rs
@@ -72,6 +72,12 @@ pub trait StateReader: Send + Sync {
         &self,
         network_id: &NetworkId,
     ) -> Result<Option<(NetworkId, SettledCertificate)>, Error>;
+
+    /// Get the local network state.
+    fn read_local_network_state(
+        &self,
+        network_id: NetworkId,
+    ) -> Result<Option<LocalNetworkStateData>, Error>;
 }
 
 pub trait PerEpochReader: Send + Sync {
@@ -94,12 +100,4 @@ pub trait PerEpochReader: Send + Sync {
         &self,
         network_id: NetworkId,
     ) -> Result<Option<Height>, Error>;
-}
-
-pub trait LocalNetworkStateReader: Send + Sync {
-    /// Get the local network state.
-    fn read_local_network_state(
-        &self,
-        network_id: NetworkId,
-    ) -> Result<Option<LocalNetworkStateData>, Error>;
 }
