+++ crates/agglayer-storage/src/stores/mod.rs
@@ -2,17 +2,14 @@ mod interfaces;
 
 pub use interfaces::{
     reader::{
-        EpochStoreReader, LocalNetworkStateReader, MetadataReader, PendingCertificateReader,
-        PerEpochReader, StateReader,
+        EpochStoreReader, MetadataReader, PendingCertificateReader, PerEpochReader, StateReader,
     },
     writer::{
-        EpochStoreWriter, LocalNetworkStateWriter, MetadataWriter, PendingCertificateWriter,
-        PerEpochWriter, StateWriter,
+        EpochStoreWriter, MetadataWriter, PendingCertificateWriter, PerEpochWriter, StateWriter,
     },
 };
 
 pub mod epochs;
-pub mod local_network_state;
 pub mod pending;
 pub mod per_epoch;
 pub mod state;
