+++ crates/agglayer-storage/src/tests/mocks.rs
@@ -1,11 +1,9 @@
 mod epochs_store;
-mod network_state_store;
 mod pending_store;
 mod per_epoch_store;
 mod state_store;
 
 pub use epochs_store::MockEpochsStore;
-pub use network_state_store::MockLocalNetworkStateStore;
 pub use pending_store::MockPendingStore;
 pub use per_epoch_store::MockPerEpochStore;
 pub use state_store::MockStateStore;
